+++ src/main/java/com/naitei/group3/movie_ticket_booking_system/config/AppConfig.java
@@ -1,0 +1,4 @@
+package com.naitei.group3.movie_ticket_booking_system.config;
+
+public class AppConfig {
+}
