+++ src/main/java/com/naitei/group3/movie_ticket_booking_system/config/SecurityConfig.java
@@ -1,0 +1,33 @@
+package com.naitei.group3.movie_ticket_booking_system.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+
+import static org.springframework.security.config.Customizer.withDefaults;
+
+@Configuration
+@EnableMethodSecurity(prePostEnabled = true)
+public class SecurityConfig {
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        // Phân quyền
+        http
+                .authorizeHttpRequests(auth -> auth
+                        .requestMatchers("/", "/public/**").permitAll()
+                        .requestMatchers("/swagger-ui/**", "/v3/api-docs/**", "/swagger-ui.html").permitAll()
+                        .requestMatchers("/admin/**").hasRole("ADMIN")
+                        .requestMatchers("/user/**").hasAnyRole("USER", "ADMIN")
+                        .anyRequest().authenticated()
+                )
+                // Xác thực
+                .formLogin(withDefaults())
+                .httpBasic(withDefaults());
+        return http.build();
+    }
+}
+
