+++ src/main/java/com/naitei/group3/movie_ticket_booking_system/entity/BookingSeat.java
@@ -1,0 +1,26 @@
+package com.naitei.group3.movie_ticket_booking_system.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Table(name = "bookingseats")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class BookingSeat {
+    @EmbeddedId
+    private BookingSeatId id;
+
+    @MapsId("bookingId")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "booking_id")
+    private Booking booking;
+
+    @MapsId("seatId")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "seat_id")
+    private Seat seat;
+}
