+++ src/main/java/com/naitei/group3/movie_ticket_booking_system/entity/PaymentTransaction.java
@@ -1,0 +1,26 @@
+package com.naitei.group3.movie_ticket_booking_system.entity;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.math.BigDecimal;
+
+@Entity
+@Table(name = "paymenttransactions")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PaymentTransaction {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "booking_id")
+    private Booking booking;
+
+    private BigDecimal amount;
+
+    private Integer status;
+}
