+++ src/main/java/com/naitei/group3/movie_ticket_booking_system/exception/AuthenticationFailedEx.java
@@ -1,0 +1,7 @@
+package com.naitei.group3.movie_ticket_booking_system.exception;
+
+public class AuthenticationFailedEx extends RuntimeException {
+    public AuthenticationFailedEx(String message) {
+        super(message);
+    }
+}
