--- src/main/java/sun/naitei/group3/movie_ticket_booking_system/MovieTicketBookingSystemApplication.java
@@ -1,13 +1,0 @@
-package sun.naitei.group3.movie_ticket_booking_system;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MovieTicketBookingSystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MovieTicketBookingSystemApplication.class, args);
-	}
-
-}
