--- src/main/java/sun/naitei/group3/movie_ticket_booking_system/config/AppConfig.java
@@ -1,4 +1,0 @@
-package sun.naitei.group3.movie_ticket_booking_system.config;
-
-public class AppConfig {
-}
