--- src/main/java/sun/naitei/group3/movie_ticket_booking_system/entity/BookingSeat.java
@@ -1,26 +1,0 @@
-package sun.naitei.group3.movie_ticket_booking_system.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "bookingseats")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class BookingSeat {
-    @EmbeddedId
-    private BookingSeatId id;
-
-    @MapsId("bookingId")
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "booking_id")
-    private Booking booking;
-
-    @MapsId("seatId")
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "seat_id")
-    private Seat seat;
-}
