--- src/main/java/sun/naitei/group3/movie_ticket_booking_system/entity/BookingSeatId.java
@@ -1,30 +1,0 @@
-package sun.naitei.group3.movie_ticket_booking_system.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-@Embeddable
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class BookingSeatId implements Serializable {
-    private Long bookingId;
-    private Long seatId;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        BookingSeatId that = (BookingSeatId) o;
-        return Objects.equals(bookingId, that.bookingId) && Objects.equals(seatId, that.seatId);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(bookingId, seatId);
-    }
-}
