--- src/main/java/sun/naitei/group3/movie_ticket_booking_system/entity/Genre.java
@@ -1,28 +1,0 @@
-package sun.naitei.group3.movie_ticket_booking_system.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-import java.util.HashSet;
-import java.util.Set;
-
-@Entity
-@Table(name = "genres")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Genre {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    @ManyToMany(mappedBy = "genres", fetch = FetchType.LAZY)
-    @JsonIgnore
-    private Set<Movie> movies= new HashSet<>(); ;
-}
