--- src/main/java/sun/naitei/group3/movie_ticket_booking_system/entity/Permission.java
@@ -1,30 +1,0 @@
-package sun.naitei.group3.movie_ticket_booking_system.entity;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import java.util.HashSet;
-import java.util.Set;
-
-@Entity
-@Table(name = "permissions")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Permission {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    @Column(columnDefinition = "TEXT")
-    private String description;
-
-    @ManyToMany(mappedBy = "permissions", fetch = FetchType.LAZY)
-    @JsonIgnore
-    private Set<Role> roles = new HashSet<>();
-}
