--- src/main/java/sun/naitei/group3/movie_ticket_booking_system/entity/RatingId.java
@@ -1,29 +1,0 @@
-package sun.naitei.group3.movie_ticket_booking_system.entity;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-@Embeddable
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class RatingId implements Serializable {
-    private Long userId;
-    private Long movieId;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        RatingId that = (RatingId) o;
-        return Objects.equals(userId, that.userId) && Objects.equals(movieId, that.movieId);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(userId, movieId);
-    }
-}
