--- src/main/java/sun/naitei/group3/movie_ticket_booking_system/exception/AuthenticationFailedEx.java
@@ -1,7 +1,0 @@
-package sun.naitei.group3.movie_ticket_booking_system.exception;
-
-public class AuthenticationFailedEx extends RuntimeException {
-    public AuthenticationFailedEx(String message) {
-        super(message);
-    }
-}
