+++ src/test/java/com/naitei/group3/movie_ticket_booking_system/MovieTicketBookingSystemApplicationTests.java
@@ -1,0 +1,13 @@
+package com.naitei.group3.movie_ticket_booking_system;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MovieTicketBookingSystemApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
